+++ Assets/ReflexPlus/Runtime/Core/Binding.cs
@@ -10,14 +10,23 @@ namespace ReflexPlus.Core
         {
         }
 
+        private Binding(IResolver resolver, Type contract, object key = null)
+        {
+            Resolver = resolver;
+            Contract = contract;
+            Key = key;
+        }
+
         private Binding(IResolver resolver, Type[] contracts, object key = null)
         {
             Resolver = resolver;
             Contracts = contracts;
             Key = key;
         }
-        
+
         public IResolver Resolver { get; }
+
+        public Type Contract { get; }
 
         public Type[] Contracts { get; }
 
@@ -25,8 +34,19 @@ namespace ReflexPlus.Core
 
         public static Binding Validated(IResolver resolver,
             Type concrete,
-            object key = null,
-            params Type[] contracts)
+            Type contract,
+            object key = null)
+        {
+            if (!contract.IsAssignableFrom(concrete))
+                throw new ContractDefinitionException(concrete, contract);
+            
+            return new Binding(resolver, contract, key);
+        }
+
+        public static Binding Validated(IResolver resolver,
+            Type concrete,
+            Type[] contracts,
+            object key = null)
         {
             foreach (var contract in contracts)
             {